+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [8.2.0-rc.1] - 2026-09-09
 
 ### Changed
@@ -1470,6 +1472,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-user/compare/staging...dev
 [8.2.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v8.1.0
 [8.1.0]: https://https://github.com/internetguru/laravel-user/compare/v8.0.0...v8.1.0
 [8.1.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v8.0.0